Implemented the functionality to obtain metrics from the website.
Implemented the functionality to perform metrics on downloaded data.

Implemented the functionality to perform metrics on the data of experiments performed locally.

Implemented the functionality to show the outliers of a parser.

The files with the data downloaded from the website and the results of the local experiments are included.